Simplify naming of dashboard components
