+++ extension/src/dashboard/Dashboard.scss
