+++ extension/src/dashboard/DashboardContent.test.tsx
@@ -2,11 +2,11 @@ import { render, screen, within } from '@testing-library/react'
 import * as React from 'react'
 import '@testing-library/jest-dom'
 import '@testing-library/jest-dom/jest-globals'
-import { MainContent, MainContentProps } from './MainContent'
-import { data } from './MainContent.test.data'
+import { data } from './DashboardContent.test.data'
+import { DashboardContent, DashboardContentProps } from './DashboardContent'
 
 test('Render and check layout', async () => {
-    render(<MainContent {...(data as unknown as MainContentProps)} />)
+    render(<DashboardContent {...(data as unknown as DashboardContentProps)} />)
 
     await screen.findByRole('heading')
 
@@ -17,7 +17,7 @@ test('Render and check layout', async () => {
 })
 
 test('Check all pipelines are included', async () => {
-    render(<MainContent {...(data as unknown as MainContentProps)} />)
+    render(<DashboardContent {...(data as unknown as DashboardContentProps)} />)
 
     const rows = screen.getAllByRole('row')
     expect(rows).toHaveLength(8) // includes header
@@ -33,7 +33,7 @@ test('Check all pipelines are included', async () => {
 })
 
 test('Check all environments are included', async () => {
-    render(<MainContent {...(data as unknown as MainContentProps)} />)
+    render(<DashboardContent {...(data as unknown as DashboardContentProps)} />)
     for (const env of data.environments) {
         const column = screen.getByText(env.name)
         expect(column).toBeInTheDocument()
